Add a ApplicationScope to create on ApplicationComponent.java, which would make it possible that all provided dependencies will be a singleton in the application level if it has been annotated with @ApplicationScope.
In DataBindModule.java, we annotated PremiumRepository as ApplicationScope.
This will mean that PremiumRepository will server as a singleton.
And its construct method will be only called once. You cold verify it by the huge log printed in the console.

In DataProviderModule.java, UserRepository was not annotated as Application Scope, it is not a singleton. So it will be created as many times as it injected.